// types/nurseEmploymentInfo.d.ts
export interface NurseEmployment {
  Address: string;
  alley: string;
  birthday: string;
  bornCity: string;
  city: string;
  education:string;
  fatherName: string;
  houseNumber?:string;
  name: string;
  nationalCode: string;
  nationalNumber: string;
  neighborHood: string;
  phoneNumber: string;
  postalCode: string;
  province: string;
  street: string;
  unit?: string;
}